[Refactor] Remove AssetManager layer as unable to properly test send function
Summary:
In D42, the `assetManager.send` function was not actually tested (`token.transfer` was tested instead), and in fact the `assetManager.address = owner` hack does not work. There is no problem functionality-wise, but in order to test AssetManager independently, we have to invoke the `send` function as AssetManager itself, so that `msg.sender` will be the address of AssetManager, which gets passed to the low-level token contract to update the balance.

The problem is we can't set the caller of `send` function to the address of AssetManager instance because it's not an address already deployed to the network. It seems impossible to simulate the condition after some research. Therefore, I moved the AssetManager logic to Core instead. Tests are also updated to reflect the change.

Test Plan: Tests included.

Reviewers: tyr034

Reviewed By: tyr034

Differential Revision: http://47.244.8.26/D47