[Refactor] Remove client dependency on TokenFactory
Summary:
This is a refactoring diff that fixes T41 and T55. The diff is quite large but necessary to remove client dependency on `TokenFactory`. The client will be reading token address from a generated file `network.json`, which includes token information for various network, whenever we run script that deploy tokens. Depending on the network type web3 detected on client-side, it reads corresponding token address without relying on calling `TokenFactory`.

Here is the change summary:

- Remove migration files for `TokenFactory` and `WETH9`.
- Deploy ERC20 and WETH contracts directly inside scripts, instead of deploying TokenFactory and create token through it.
- Update token info in `network.json` after token deployment.
- Remove `TokenFactory` dependency from client `Web3Service`.
- Client method `TokenService.getTokenAddress` now detects network type and read corresponding token address from `network.json`.
- Refactor scripts and tests that depends on `TokenFactory`.

To help with the review, I will list the essential changed files here:

- `client/src/stores/services/TokenService.ts`
- `client/src/stores/services/Web3Service.ts`
- `scripts/javascript/deployTokens.js`
- `scripts/javascript/setupEnvironment.js`

@ZhangRGK once this is landed, don't forget to update your `config.js` and run `setupEnvironment` again.

Reviewers: ZhangRGK

Reviewed By: ZhangRGK

Maniphest Tasks: T41, T55

Differential Revision: https://phabricator.bluestone.live/D247