[Contract] Fix repay to wrong pool
Summary:
Currently we repay to the pool according to the poolIndex (stored in Loan record), which is wrong. Instead, we should record poolId in Loan record and repay to the pool with the same poolId instead. To illustrate it with a 30-day pool group with poolIds from 0 to 29:

1st day: [0, 1, 2, ..., 27, 28, 29]
2nd day: [1, 2, 3, ..., 28, 29, 0]
3rd day: [2, 3, 4, ..., 29, 0, 1]

Our current behavior for a 30-day loan and repays after two day is: we loan from the last pool (id: 29), store poolIndex 29 in loan record. After two days, we get poolId at poolIndex 29, which is 1, and we repay to pool with id 1. What we should repay to is pool with id 29 at index 27.

This change stores poolId 29 in loan record instead, and we check for loan amount in record when we reach poolIndex 27, we found loan amount > 0, so we repay back to pool with id 29 at index 27.

Test is added to ensure it repays to the correct pool we loaned from.

Fix T107.

Reviewers: ZhangRGK, tyr034

Reviewed By: ZhangRGK, tyr034

Maniphest Tasks: T107

Differential Revision: https://phabricator.bluestone.live/D311