[Contract]: store freed collateral by correct key.
Summary:
Fix 1 bugs:

The msg.sender value will be caller contract during method calling between 2 contract,

and I found tx.origin was deprecated.

so we need to pass the user as a parameter to account manager

---

redesign how to use freed collateral in user actions:

I think pass the freed collateral amount as a parameter from client is not safety.

so I changed it to only pass the flag that use freed collateral or not.

Reviewers: desmonding, tyr034, lambda

Reviewed By: desmonding, tyr034

Differential Revision: https://phabricator.bluestone.live/D320