[Contract] Drop asset list argument for deposit maturity update
Summary:
Previously we require admin to specify a list of asset for deposit maturity update since we didn't have a data structure to store deposit assets. Now, it's much safer for the contract to update all deposit assets we store internally instead of relying on user input, which is error-prone.

- Remove `assetList` argument from `updateAllDepositMaturity`.
- Add data structure to store enabled and disabled deposit assets.

Reviewers: ZhangRGK, lambda, tyr034

Reviewed By: ZhangRGK

Differential Revision: https://phabricator.bluestone.live/D342