[Contract] Implement basic integration between library and Protocol
Summary:
This is the very first step to create basic structure of contract and test code by integrating a simple method in Configuration lib to Protocol contract.

Note: library name is prefixed with `_` to avoid conflict with existing contract. It will be removed after refactor is complete.

Closes T136.

Reviewers: ZhangRGK, lambda, tyr034

Reviewed By: ZhangRGK

Maniphest Tasks: T136

Differential Revision: https://phabricator.bluestone.live/D354