Fix soldCollateralAmount calculation in liquidateLoan
Summary:
When calculating collateral amount to be sold to liquidator in `liquidateLoan` function and loan token has less than 18 decimals, it will result in a collateral amount less than expected as seen in T384. The fix is to make sure the value of loan tokens (amount * price) is in the scale of 1e18 (same as the collateral token price) when doing the calculation.

Fix T384.

Reviewers: yangwenyi, lu0ty, tyr034

Reviewed By: yangwenyi

Maniphest Tasks: T384

Differential Revision: https://phabricator.bluestone.live/D759