[Contract] Store matched pool ids for loan records so that we could loop though matched pools to repay
Summary:
**Problem**
When repaying a loan, we loop through all active pools and repay to the pool if `loanAmountByLoanIdAndPoolId[loanId][poolId]` is greater than 0. But this could be an issue if the loan is repaid after one of the matched pools matures. Matured pools won't be checked by the loop.

**Solution**
Store matched pool ids for loan records in PoolGroup so that we could loop though all matched pools (either matured or not) to repay.

Close T417

Reviewers: yangwenyi, tyr034

Reviewed By: yangwenyi, tyr034

Subscribers: lu0ty

Tags: #bluestone, #contract, #bug

Maniphest Tasks: T417

Differential Revision: https://phabricator.bluestone.live/D782