[Contract] Fixed balance cap checking bug for depositing ETH.
Summary:
The origin issue described in T444 is actually not an issue. Because `address(this).balance` already includes the actual deposit amount, the following require check still works even if `depositParameters.depositAmount` is 0.
```
require(
    currTokenBalance.add(depositParameters.depositAmount) <=
    configuration.balanceCapByToken[depositParameters.tokenAddress],
    'DepositManager: token balance cap exceeded'
);
```
However, another issue comes up. If `depositParameters.depositAmount` is greater than 0, the above require check may throw error even if the token cap is not reached. For example, when the cap is 100, and both `depositParameters.depositAmount` and `msg.value` are 100, the transaction will fail because `currTokenBalance.add(depositParameters.depositAmount)` = 200 > 100.

The solution here is to introduce a new variable `tokenBalanceAfterTx` for the token balance after the transaction succeeds and require tokenBalanceAfterTx <= token cap.

Close T444

Reviewers: tyr034, yangwenyi

Reviewed By: yangwenyi

Tags: #bluestone, #contract, #bug

Maniphest Tasks: T444

Differential Revision: https://phabricator.bluestone.live/D802