[Contract][Part 2] Require xxxAmount in function parameters to be equal to msg.value for ETH-related transactions in LoanManager.sol
Summary:
In `DepositManager.sol`, `loan()`, `repayLoan()`, `liquidateLoan()` and `addCollateral()` are affected by this refactoring. `subtractCollateral()` is not changed because it's not payable and require msg.value to be zero.

There are many changes in the `LoanManager.test.js`. Many of them are previously existing test cases, but with different indentations. I'll mark all new test cases in comments.

Reviewers: tyr034, yangwenyi

Reviewed By: yangwenyi

Tags: #bluestone, #contract

Maniphest Tasks: T446

Differential Revision: https://phabricator.bluestone.live/D825